+++ src/math/p_popcount.c
@@ -1,5 +1,5 @@
 #include <pal.h>
-
+#include <stdint.h>
 /**
  *
  * Counts the number of bits set in 'a'.
@@ -17,7 +17,7 @@
  * @return      None
  *
  */
-void p_popcount_u32(unsigned int *a, unsigned int *c, int n, int p, p_team_t team)
+void p_popcount_u32(uint32_t *a, uint32_t *c, int n, int p, p_team_t team)
 {
     static const unsigned int A[] = {0x55555555, 0x33333333,
                                      0x0f0f0f0f, 0x01010101};
@@ -37,7 +37,7 @@ void p_popcount_u32(unsigned int *a, unsigned int *c, int n, int p, p_team_t tea
     }
 }
 
-void p_popcount_u64(unsigned int *a, unsigned int *c, int n, int p, p_team_t team)
+void p_popcount_u64(uint64_t *a, uint64_t *c, int n, int p, p_team_t team)
 {
     /* Implement me */
 }
